% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_charts.R
\name{available_charts}
\alias{available_charts}
\title{List of available charts in fusionchartsR}
\usage{
available_charts()
}
\description{
List of available charts in fusionchartsR
}
\examples{
library(fusionchartsR)
available_charts()

}
