% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionTheme.R
\name{fusionTheme}
\alias{fusionTheme}
\title{Adding FusionCharts theme}
\usage{
fusionTheme(
  fusionPlot,
  theme = c("fusion", "gammel", "candy", "zune", "ocean", "carbon", "umber")
)
}
\arguments{
\item{fusionPlot}{fusionPlot object got by \code{fusionPlot()}}

\item{theme}{Chart theme}
}
\description{
\url{https://www.fusioncharts.com/dev/themes/introduction-to-themes}
}
\examples{
library(fusionchartsR)
df <- data.frame(label = c("Venezuela", "Saudi", "Canada", "Russia"), value = c(290, 260,180, 115))
df \%>\%
fusionPlot(x = "label", y = "value", type = "pie2d") \%>\%
fusionTheme(theme = "gammel")


}
