% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfusionplot.R
\name{ggfusionplot}
\alias{ggfusionplot}
\title{Convert ggplot2 to fusionchartsR}
\usage{
ggfusionplot(object)
}
\arguments{
\item{object}{a ggplot object}
}
\description{
This function converts a [ggplot2::ggplot()] object to a fusioncharstR object.
}
\examples{
library(fusionchartsR)
library(ggplot2)

object <- ggplot(data = mtcars, aes(x = wt, y = mpg)) +
  geom_point()
 
ggfusionplot(object)
}
