% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfusionPlot.R
\name{ggfusionPlot}
\alias{ggfusionPlot}
\title{Convert ggplot2 to fusionchartsR}
\usage{
ggfusionPlot(object)
}
\arguments{
\item{object}{a ggplot object}
}
\description{
This function converts a [ggplot2::ggplot()] object to a fusioncharstR object.
}
\examples{
library(fusionchartsR)
library(ggplot2)

object <- ggplot(data = mtcars, aes(x = wt, y = mpg)) +
  geom_point()

ggfusionPlot(object)
}
